separated to new and legacy scripts
